范围型技能释放，包括释放范围选择特效与范围选择逻辑，待测试

+++ Src/Lib/Common/Data/SkillDefine.cs
@@ -15,7 +15,7 @@ namespace Common.Data
         public SkillType Type { get; set; }
         public int UnlockLevel { get; set; }
         public TargetType CastTarget { get; set; }
-        public float CastRange { get; set; }
+        public int CastRange { get; set; }
         public float CastTime { get; set; }
         public float CD { get; set; }
         public int MPCost { get; set; }
